Canonicalize test units with @safetestset for isolation
Convert the plain-@testset test units to @safetestset so each
independent unit runs in its own module (isolation + world-age
safety), matching the canonical SciML/OrdinaryDiffEq structure.

- runtests.jl: extract the loose top-level solver test block (N-body
  setup, analytical convergence, backward-integration) into a new
  self-contained test/main_tests.jl and wrap it as
  `@safetestset "IRKGL16 Solver" begin include("main_tests.jl") end`.
  The "Allocation Tests" and "Quality Assurance" includes become
  @safetestset. alloc_tests.jl already carries its own imports.
- qa/qa.jl: convert the Aqua and JET units to @safetestset, each with
  its own self-contained imports; @test_broken preserved.
- Add SafeTestsets to the test deps (root Project.toml
  [extras]/[targets].test/[compat] and test/qa/Project.toml).

Behavior-preserving: same tests, same assertions, same GROUP dispatch.
Verified locally with julia 1.11: GROUP=Core (IRKGL16 Solver 4/4,
Allocation Tests 6/6) and GROUP=QA (8 pass, 1 broken) both pass.
