+++ Project.toml
@@ -20,6 +20,7 @@ ODEProblemLibrary = "1"
 Parameters = "0.12, 0.13"
 Reexport = "1.2.2"
 SIMD = "3.5.0"
+SafeTestsets = "0.1, 1"
 SciMLBase = "3.1"
 SciMLLogging = "1.10.1, 2"
 Test = "1"
@@ -29,7 +30,8 @@ julia = "1.10"
 AllocCheck = "9b6a8646-10ed-4001-bbdc-1d2f46dfbb1a"
 DiffEqDevTools = "f3b72e0c-5b89-59e1-b016-84e28bfd966d"
 ODEProblemLibrary = "fdc4e326-1af4-4b90-96e7-779fcce2daa5"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["AllocCheck", "DiffEqDevTools", "ODEProblemLibrary", "Test"]
+test = ["AllocCheck", "DiffEqDevTools", "ODEProblemLibrary", "SafeTestsets", "Test"]
